+++ src/prefect/agent/local/agent.py
@@ -137,7 +137,7 @@ class LocalAgent(Agent):
             "PREFECT__CLOUD__API": config.cloud.api,
             "PREFECT__CLOUD__AGENT__LABELS": str(self.labels),
             "PREFECT__CONTEXT__FLOW_RUN_ID": flow_run.id,  # type: ignore
-            "PREFECT__CONTEXT__FLOW_RUN_NAME": '"{}"'.format(flow_run.name),  # type: ignore
+            "PREFECT__CONTEXT__FLOW_RUN_NAME": flow_run.name,  # type: ignore
             "PREFECT__CLOUD__USE_LOCAL_SECRETS": "false",
             "PREFECT__LOGGING__LOG_TO_CLOUD": "true",
             "PREFECT__LOGGING__LEVEL": "DEBUG",