+++ tests/agent/test_nomad_agent.py
@@ -101,5 +101,5 @@ def test_nomad_agent_replace_yaml(runner_token):
         assert env["PREFECT__CLOUD__API"] == "https://api.prefect.io"
         assert env["PREFECT__CLOUD__AGENT__AUTH_TOKEN"] == "token"
         assert env["PREFECT__CONTEXT__FLOW_RUN_ID"] == "id"
-        assert env["PREFECT__CONTEXT__FLOW_RUN_NAME"] == '"name"'
+        assert env["PREFECT__CONTEXT__FLOW_RUN_NAME"] == "name"
         assert env["PREFECT__CONTEXT__NAMESPACE"] == "default"